* Source/gsc/GSStreamContext.m: * Source/gsc/GSContext.m: * Source/gsc/GSGState.m: * Headers/gsc/GSGStateOps.h: Change type for -GSSetFillColor: and -GSSetStrokeColor: from float to CGFloat
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/back/trunk@34071 72102866-910b-0410-8b05-ffd578937521

+++ Headers/gsc/GSGStateOps.h
@@ -46,8 +46,8 @@
 
 - (void) GSSetFillColorspace: (void *)spaceref;
 - (void) GSSetStrokeColorspace: (void *)spaceref;
-- (void) GSSetFillColor: (const float *)values;
-- (void) GSSetStrokeColor: (const float *)values;
+- (void) GSSetFillColor: (const CGFloat *)values;
+- (void) GSSetStrokeColor: (const CGFloat *)values;
 
 /* ----------------------------------------------------------------------- */
 /* Text operations */
